fix somew llvm/webgl warnings.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@4757 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/client/pr_csqc.c
@@ -855,7 +855,7 @@ static void QCBUILTIN PF_R_DynamicLight_Get(pubprogfuncs_t *prinst, struct globa
 	enum lightfield_e field = G_FLOAT(OFS_PARM1);
 	if (lno >= rtlights_max)
 	{
-		if (field == -1)
+		if ((int)field == -1)
 			G_FLOAT(OFS_RETURN) = rtlights_max;
 		else
 			G_INT(OFS_RETURN) = 0;
